Fix runtime view controller warning in console
The existing code was trying to present a view controller from the Dashboard view controller at -[viewDidLoad] which is too early as the Dashboard view controller is detached at this point. Moving this code to -[didMoveToParentViewController:] resolves the issue as the view controller has been added into the navigation stack by then. Note that this also causes the code to run every time the user returns to this screen, but I think this is completely reasonable. Will need to revisit this though because there is a small flicker when first launching the app and the login screen being presented over the dashboard.